pipeline training with the dataset output
